+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@
 - fix: failed to load json/data content in api reference
 - fix: failed to convert postman collection containing multipart/form-data requests to pytest
 - fix: only get the first parameter in referenced testcase
+- refactor: simplify testcase converter
 
 **python version**